# java-vertx-server/src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "\nA set of APIs based on the LSO Reference Architecture for\nServiceability (Address Validation, Site Queries, Product Offering Qualification) |\nQuoting | Product Inventory | Ordering | Trouble-ticketing Billing | Contract & Catalog\n\n",
    "version" : "3.0.0",
    "title" : "MEF LSO Sonata",
    "x-logo" : {
      "backgroundColor" : "#FFFFFF",
      "url" : "/redoc/logo.png"
    }
  },
  "host" : "serverRoot",
  "basePath" : "/api/mef",
...
  "schemes" : [ "https" ],
  "produces" : [ "application/json;charset=utf-8" ],
  "security" : [ {
    "bearer" : [ ]
  } ],
  "paths" : {
...
    "/geographicAddressManagement/v3/geographicAddressValidation" : {
      "post" : {
        "tags" : [ "GeographicAddressValidation" ],
        "summary" : "validate a Geographic Address",
        "description" : "The Buyer sends Address information known to the Buyer to the Seller.  The Seller re-sponds with a list of Addresses known to the Seller that likely match the Address infor-mation sent by the Buyer.  For each Address returned, the Seller generally also provides an Address Identifier, which uniquely identifies this Address within the Seller.",
        "operationId" : "geographicAddressValidationCreate",
        "consumes" : [ "application/json;charset=utf-8" ],
        "produces" : [ "application/json;charset=utf-8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "addressValidationRequest",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/GeographicAddressValidation_Create"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Created",
            "schema" : {
              "$ref" : "#/definitions/GeographicAddressValidation"
            }
          },
          "400" : {
            "description" : "Bad Request\n\nList of supported error codes:\n- 20: Invalid URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string parameter value",
            "schema" : {
              "$ref" : "#/definitions/ErrorRepresentation"
            }
          },
          "401" : {
            "description" : "Unauthorized\n\nList of supported error codes:\n- 40: Missing credentials\n- 41: Invalid credentials\n- 42: Expired credentials",
            "schema" : {
              "$ref" : "#/definitions/ErrorRepresentation"
            }
          },
          "403" : {
            "description" : "Forbidden\n\nList of supported error codes:\n- 50: Access denied\n- 51: Forbidden requester\n- 52: Forbidden user\n- 53: Too many requests",
            "schema" : {
              "$ref" : "#/definitions/ErrorRepresentation"
            }
          },
          "404" : {
            "description" : "Not Found\n\nList of supported error codes:\n- 60: Resource not found",
            "schema" : {
              "$ref" : "#/definitions/ErrorRepresentation"
            }
          },
          "405" : {
            "description" : "Method Not Allowed\n\nList of supported error codes:\n- 61: Method not allowed",
            "schema" : {
              "$ref" : "#/definitions/ErrorRepresentation"
            }
          },
          "422" : {
            "description" : "Unprocessable entity\n\nFunctional error\n\n\n\n\n\n - code: 100\nmessage: Too many addresses match - please re-try with more attribute values restriction\ndescription:",
            "schema" : {
              "$ref" : "#/definitions/ErrorRepresentation"
            }
          },
          "500" : {
            "description" : "Internal Server Error\n\nList of supported error codes:\n- 1: Internal error",
            "schema" : {
              "$ref" : "#/definitions/ErrorRepresentation"
            }
          },
          "503" : {
            "description" : "Service Unavailable\n\n",
            "schema" : {
              "$ref" : "#/definitions/ErrorRepresentation"
            }
          }
        },
        "security" : [ {
          "bearer" : [ "customer", "admin" ]
        } ],
        "x-contentType" : "application/json;charset=utf-8",
        "x-accepts" : "application/json;charset=utf-8"
      }
    },
...
  },
  "securityDefinitions" : {
    "bearer" : {
      "type" : "oauth2",
      "authorizationUrl" : "https://dummy.oauth.net/auth",
      "tokenUrl" : "https://dumy.oauth.net/token",
      "flow" : "accessCode",
      "scopes" : {
        "customer" : "scope of registered customers",
        "admin" : "scope of admin customers"
      }
    }
  },
  "definitions" : {
...
    "ErrorRepresentation" : {
      "type" : "object",
      "required" : [ "code", "reason" ],
      "properties" : {
        "@schemaLocation" : {
          "type" : "string",
          "description" : "it provides a link to the schema describing a REST resource."
        },
        "@type" : {
          "type" : "string",
          "description" : "The class type of a REST resource."
        },
        "code" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Application related code (as defined in the API or from a common list)"
        },
        "message" : {
          "type" : "string",
          "description" : "Text that provide more details and corrective actions related to the error. This can be shown to a client user."
        },
        "reason" : {
          "type" : "string",
          "description" : "Text that explains the reason for error. This can be shown to a client user."
        },
        "referenceError" : {
          "type" : "string",
          "description" : "url pointing to documentation describing the error"
        },
        "status" : {
          "type" : "string",
          "description" : "http error code extension like 400-2"
        }
      },
      "description" : "Class used to describe API response error"
    },
...
    "FieldedAddress" : {
      "type" : "object",
      "properties" : {
        "city" : {
          "type" : "string",
          "description" : "City that the address is in"
        },
        "country" : {
          "type" : "string",
          "description" : "Country that the address is in"
        },
        "geographicSubAddress" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/GeographicSubAddress"
          }
        },
        "id" : {
          "type" : "string",
          "description" : "Unique identifier of the address"
        },
        "locality" : {
          "type" : "string",
          "description" : "\"An area of defined or undefined boundaries within a local authority or other legislatively defined area, usually rural or semi-rural in nature.\" [ANZLIC-STREET], or a suburb \"a bounded locality within a city, town or shire principally of urban character \" [ANZLICSTREET]."
        },
        "postCodeExtension" : {
          "type" : "string",
          "description" : "The four-digit extension on an American postal code, what comes after the hyphen when specified."
        },
        "postcode" : {
          "type" : "string",
          "description" : "Descriptor for a postal delivery area, used to speed and simplify the delivery of mail (also known as zipcode)"
        },
        "stateOrProvince" : {
          "type" : "string",
          "description" : "The State or Province that the address is in"
        },
        "streetName" : {
          "type" : "string",
          "description" : "Name of the street or other street type"
        },
        "streetNr" : {
          "type" : "string",
          "description" : "Number identifying a specific property on a public street. It may be combined with streetNrLast for ranged addresses"
        },
        "streetNrLast" : {
          "type" : "string",
          "description" : "Last number in a range of street numbers allocated to a property"
        },
        "streetNrLastSuffix" : {
          "type" : "string",
          "description" : "Last street number suffix for a ranged address"
        },
        "streetNrSuffix" : {
          "type" : "string",
          "description" : "The first street number suffix"
        },
        "streetSuffix" : {
          "type" : "string",
          "description" : "A modifier denoting a relative direction"
        },
        "streetType" : {
          "type" : "string",
          "description" : "Alley, avenue, boulevard, brae, crescent, drive, highway, lane, terrace, parade, place, tarn, way, wharf"
        }
      },
      "description" : "A type of Address that has a discrete field and value for each type of boundary or identifier down to the lowest level of detail. For example “street number” is one field, “street name” is another field, etc."
    },
    "FieldedAddressRequest" : {
      "type" : "object",
      "properties" : {
        "city" : {
          "type" : "string",
          "description" : "City that the address is in"
        },
        "country" : {
          "type" : "string",
          "description" : "Country that the address is in"
        },
        "geographicSubAddress" : {
          "$ref" : "#/definitions/GeographicSubAddressRequest"
        },
        "locality" : {
          "type" : "string",
          "description" : "\"An area of defined or undefined boundaries within a local authority or other legislatively defined area, usually rural or semi-rural in nature.\" [ANZLIC-STREET], or a suburb \"a bounded locality within a city, town or shire principally of urban character \" [ANZLICSTREET]."
        },
        "postCodeExtension" : {
          "type" : "string",
          "description" : "The four-digit extension on an American postal code, what comes after the hyphen when specified."
        },
        "postcode" : {
          "type" : "string",
          "description" : "Descriptor for a postal delivery area, used to speed and simplify the delivery of mail (also known as zipcode)"
        },
        "stateOrProvince" : {
          "type" : "string",
          "description" : "The State or Province that the address is in"
        },
        "streetName" : {
          "type" : "string",
          "description" : "Name of the street or other street type"
        },
        "streetNr" : {
          "type" : "string",
          "description" : "Number identifying a specific property on a public street. It may be combined with streetNrLast for ranged addresses"
        },
        "streetNrLast" : {
          "type" : "string",
          "description" : "Last number in a range of street numbers allocated to a property"
        },
        "streetNrLastSuffix" : {
          "type" : "string",
          "description" : "Last street number suffix for a ranged address"
        },
        "streetNrSuffix" : {
          "type" : "string",
          "description" : "The first street number suffix"
        },
        "streetSuffix" : {
          "type" : "string",
          "description" : "A modifier denoting a relative direction"
        },
        "streetType" : {
          "type" : "string",
          "description" : "Alley, avenue, boulevard, brae, crescent, drive, highway, lane, terrace, parade, place, tarn, way, wharf"
        }
      },
      "description" : "A type of Address that has a discrete field and value for each type of boundary or identifier down to the lowest level of detail. For example “street number” is one field, “street name” is another field, etc.\nThis structure is used in the request addressValidation only"
    },
    "FormattedAddress" : {
      "type" : "object",
      "required" : [ "addrLine1" ],
      "properties" : {
        "addrLine1" : {
          "type" : "string",
          "description" : "The first address line in a formatted address"
        },
        "addrLine2" : {
          "type" : "string",
          "description" : "The second address line in a formatted address"
        },
        "city" : {
          "type" : "string",
          "description" : "City that the address is in"
        },
        "country" : {
          "type" : "string",
          "description" : "Country that the address is in"
        },
        "id" : {
          "type" : "string",
          "description" : "Unique identifier of the address"
        },
        "locality" : {
          "type" : "string",
          "description" : "\"An area of defined or undefined boundaries within a local authority or other legislatively defined area, usually rural or semi-rural in nature.\" [ANZLIC-STREET], or a suburb \"a bounded locality within a city, town or shire principally of urban character \" [ANZLICSTREET]."
        },
        "postCodeExtension" : {
          "type" : "string",
          "description" : "The four-digit extension on an American postal code, what comes after the hyphen when specified."
        },
        "postcode" : {
          "type" : "string",
          "description" : "Descriptor for a postal delivery area, used to speed and simplify the delivery of mail (also known as zipcode)"
        },
        "stateOrProvince" : {
          "type" : "string",
          "description" : "The State or Province that the address is in"
        }
      },
      "description" : "A type of Address that has discrete fields for each type of boundary or identifier with the exception of street and more specific location details, which are combined into a maximum of two strings based on local postal addressing conventions"
    },
    "FormattedAddressRequest" : {
      "type" : "object",
      "required" : [ "addrLine1" ],
      "properties" : {
        "addrLine1" : {
          "type" : "string",
          "description" : "The first address line in a formatted address"
        },
        "addrLine2" : {
          "type" : "string",
          "description" : "The second address line in a formatted address"
        },
        "city" : {
          "type" : "string",
          "description" : "City that the address is in"
        },
        "country" : {
          "type" : "string",
          "description" : "Country that the address is in"
        },
        "locality" : {
          "type" : "string",
          "description" : "\"An area of defined or undefined boundaries within a local authority or other legislatively defined area, usually rural or semi-rural in nature.\" [ANZLIC-STREET], or a suburb \"a bounded locality within a city, town or shire principally of urban character \" [ANZLICSTREET]."
        },
        "postCodeExtension" : {
          "type" : "string",
          "description" : "The four-digit extension on an American postal code, what comes after the hyphen when specified."
        },
        "postcode" : {
          "type" : "string",
          "description" : "Descriptor for a postal delivery area, used to speed and simplify the delivery of mail (also known as zipcode)"
        },
        "stateOrProvince" : {
          "type" : "string",
          "description" : "The State or Province that the address is in"
        }
      },
      "description" : "A type of Address that has discrete fields for each type of boundary or identifier with the exception of street and more specific location details, which are combined into a maximum of two strings based on local postal addressing conventions"
    },
...
    "GeographicAddressRequestBuyerInput" : {
      "type" : "object",
      "properties" : {
        "fieldedAddress" : {
          "$ref" : "#/definitions/FieldedAddressRequest"
        },
        "formattedAddress" : {
          "$ref" : "#/definitions/FormattedAddressRequest"
        }
      },
      "description" : "Structure used by the buyer to request geographic address validation"
    },
    "GeographicAddressSellerResponse" : {
      "type" : "object",
      "required" : [ "isBestMatch" ],
      "properties" : {
        "@schemaLocation" : {
          "type" : "string",
          "description" : "Technical attribute to extend this class"
        },
        "@type" : {
          "type" : "string",
          "description" : "Technical attribute to extend this class"
        },
        "allowsNewSite" : {
          "type" : "boolean",
          "description" : "This attribute specifies if a Buyer must use one of the known existing Service Sites at this location for any Products delivered to this Address.  For example, if a particular building owner mandated that all interconnects be done in a shared Public Meet-Me-Room, this attribute would be set to False for that Address."
        },
        "fieldedAddress" : {
          "$ref" : "#/definitions/FieldedAddress"
        },
        "formattedAddress" : {
          "$ref" : "#/definitions/FormattedAddress"
        },
        "hasPublicSite" : {
          "type" : "boolean",
          "description" : "This attribute specifies if that Address contains Service Sites that are public such as Meet-Me-Rooms at an interconnect location or a shared telecom room in the basement of a multi-tenant building"
        },
        "id" : {
          "type" : "string",
          "description" : "Unique identifier of the address"
        },
        "isBestMatch" : {
          "type" : "boolean",
          "description" : "For each Address returned, this attribute specifies if the Seller believes this is the best match to the Buyer specified Address.  Setting this to TRUE also represents that the Seller believes this is likely a direct match to the Buyer Specified Address."
        }
      },
      "description" : "An address is a structured textual way of describing how to find a Property in an urban area (country properties are often defined differently)"
    },
    "GeographicAddressValidation" : {
      "type" : "object",
      "required" : [ "requestedAddress" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "description" : "Unique identifier of the Address Validation (Not an address id !!)"
        },
        "requestedAddress" : {
          "$ref" : "#/definitions/GeographicAddressRequestBuyerInput"
        },
        "validationDate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date when the address validation is performed"
        },
        "validationResult" : {
          "$ref" : "#/definitions/ValidationResult"
        },
        "verifiedAddress" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/GeographicAddressSellerResponse"
          }
        }
      },
      "description" : "Resource used to support a request for a validation address : check if a geographic address described by values attributes exists)\nIf exist, id is provided\nif not, alternate address(es) could be provided"
    },
    "GeographicAddressValidation_Create" : {
      "type" : "object",
      "required" : [ "requestedAddress" ],
      "properties" : {
        "requestedAddress" : {
          "$ref" : "#/definitions/GeographicAddressRequestBuyerInput"
        }
      },
      "description" : "Structure used to describe a geographic address validation request"
    },
...
    "GeographicSubAddress" : {
      "type" : "object",
      "properties" : {
        "@schemaLocation" : {
          "type" : "string",
          "description" : "Technical attribute to extend this class"
        },
        "@type" : {
          "type" : "string",
          "description" : "Technical attribute to extend this class"
        },
        "buildingName" : {
          "type" : "string",
          "description" : "Allows for buildings that have well-known names"
        },
        "id" : {
          "type" : "string",
          "description" : "Unique Identifier of the subAddress"
        },
        "levelNumber" : {
          "type" : "string",
          "description" : "Used where a level type may be repeated e.g. BASEMENT 1, BASEMENT 2"
        },
        "levelType" : {
          "type" : "string",
          "description" : "Describes level types within a building"
        },
        "privateStreetName" : {
          "type" : "string",
          "description" : "Private streets internal to a property (e.g. a university) may have internal names that are not recorded by the land title office"
        },
        "privateStreetNumber" : {
          "type" : "string",
          "description" : "Private streets numbers internal to a private street"
        },
        "subUnit" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SubUnit"
          }
        }
      },
      "description" : "It is used for addressing within a property in an urban area (country properties are often defined differently). It may refer to a building, a building cluster, or a floor of a multistory building."
    },
    "GeographicSubAddressRequest" : {
      "type" : "object",
      "properties" : {
        "@schemaLocation" : {
          "type" : "string",
          "description" : "Technical attribute to extend this class"
        },
        "@type" : {
          "type" : "string",
          "description" : "Technical attribute to extend this class"
        },
        "buildingName" : {
          "type" : "string",
          "description" : "Allows for buildings that have well-known names"
        },
        "levelNumber" : {
          "type" : "string",
          "description" : "Used where a level type may be repeated e.g. BASEMENT 1, BASEMENT 2"
        },
        "levelType" : {
          "type" : "string",
          "description" : "Describes level types within a building"
        },
        "privateStreetName" : {
          "type" : "string",
          "description" : "Private streets internal to a property (e.g. a university) may have internal names that are not recorded by the land title office"
        },
        "privateStreetNumber" : {
          "type" : "string",
          "description" : "Private streets numbers internal to a private street"
        },
        "subUnit" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SubUnit"
          }
        }
      },
      "description" : "It is used for addressing within a property in an urban area (country properties are often defined differently). It may refer to a building, a building cluster, or a floor of a multistory building."
    },
...
    "SubUnit" : {
      "type" : "object",
      "required" : [ "subUnitIdentifier", "subUnitType" ],
      "properties" : {
        "subUnitIdentifier" : {
          "type" : "string",
          "description" : "The discriminator used for the subunit, often just a simple number but may also be a range."
        },
        "subUnitType" : {
          "type" : "string",
          "description" : "The type of subunit e.g.BERTH, FLAT, PIER, SUITE, SHOP, TOWER, UNIT, WHARF."
        }
      },
      "description" : "Sub Unit structure is used to describe several subUnits with a geographical subAddress."
    },
...
    "ValidationResult" : {
      "type" : "string",
      "description" : "Validation Result",
      "enum" : [ "success", "fails", "partial" ]
    },
...
  }
}